Bootstrapped Edge Inclusion 'Probabilities'

Compute edge inclusion 'probabilities' with a non-parametric bootstrap.

boot_eip(
  Y,
  method = "pearson",
  samples = 50,
  penalty = "atan",
  ic = "bic",
  gamma = NULL,
  lambda = NULL,
  n_lambda = 50,
  n_gamma = 50,
  unreg = FALSE,
  progress = TRUE,
  ...
)

Arguments

Y
Matrix. A matrix of dimensions n by p.

method
Character string. Which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman." Defaults to pearson.

samples
Numeric. How many boostrap samples (defaults to 500).

penalty
Character string. Which penalty should be used (defaults to "atan")?

ic
Character string. Which information criterion should be used (defaults to "bic")? The options include aic, ebic (ebic_gamma defaults to 0.5; see details), ric, or any of the generalized information criteria provided in section 5 of Kim et al. (2012) . The options are gic_1 (i.e., bic) to gic_6.

gamma
Numeric. Hyperparameter for the penalty function. Defaults to 3.7 (SCAD), 2 (MCP), 0.5 (adapt), and 0.01 otherwise with select = "lambda".

lambda
Numeric vector. Regularization parameter. Defaults to NULL that provides default values with select = "lambda" and sqrt(log(p)/n) with select = "gamma".

n_lambda
Numeric. The number of \(\lambda\)'s to be evaluated. Defaults to 50. This is disregarded if custom values are provided in lambda.

n_gamma
Numeric. The number of \(\gamma\)'s to be evaluated. Defaults to 50. This is disregarded if custom values are provided in lambda.

unreg
Logical. Should the models be refitted (or unregularized) with maximum likelihood (defaults to FALSE)? Setting to TRUE results in the approach of Foygel and Drton (2010) , but with the regularization path obtained from nonconvex regularization, as opposed to the \(\ell_1\)-penalty.

progress
Logical. Should a progress bar be included (defaults to TRUE) ?

...
Additional arguments. Currently gamma in EBIC (ic = "ebic") can be set with ebic_gamma = 1.

Value

An object of class eip

Examples

# data
Y <- GGMncv::ptsd[,1:10]

# compute eip's
boot_samps <- boot_eip(Y, samples  = 10)
#> 
#> computing eip's
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%