De-Sparsified Graphical Lasso Estimator

Compute the de-sparsified glasso estimator with the approach described in Equation 7 of Jankova and Van De Geer (2015) .

desparsify(object, ...)

Arguments

object
An object of class ggmncv

...
Currently ignored

Value

The de-sparsified estimates, including

Theta: De-sparsified precision matrix

P: De-sparsified partial correlation matrix

Details

According to Jankova and Van De Geer (2015) , the de-sparisifed estimator, \(\hat{\mathrm{\bf T}}\), is defined as

\(\hat{\mathrm{\bf T}} = 2\hat{\boldsymbol{\Theta}} - \hat{\boldsymbol{\Theta}}\hat{\mathrm{\bf R}}\hat{\boldsymbol{\Theta}},\)

where \(\hat{\boldsymbol{\Theta}}\) denotes the graphical lasso estimator of the precision matrix and \(\hat{\mathrm{\bf R}}\) is the sample correlation matrix. Further details can be found in section 2 ("Main Results") of Jankova and Van De Geer (2015) .

Note

This assumes (reasonably) Gaussian data.

References

Jankova J, Van De Geer S (2015). “Confidence intervals for high-dimensional inverse covariance estimation.” Electronic Journal of Statistics, 9(1), 1205--1229.

Examples

# data
Y <- GGMncv::Sachs

# fit model
fit <- ggmncv(cor(Y), n = nrow(Y))
#> selecting lambda
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   6%  |                                                                              |======                                                                |   8%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  12%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  16%  |                                                                              |=============                                                         |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  26%  |                                                                              |====================                                                  |  28%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  32%  |                                                                              |========================                                              |  34%  |                                                                              |=========================                                             |  36%  |                                                                              |===========================                                           |  38%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  46%  |                                                                              |==================================                                    |  48%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  52%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  56%  |                                                                              |=========================================                             |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  66%  |                                                                              |================================================                      |  68%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  72%  |                                                                              |====================================================                  |  74%  |                                                                              |=====================================================                 |  76%  |                                                                              |=======================================================               |  78%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  86%  |                                                                              |==============================================================        |  88%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  92%  |                                                                              |==================================================================    |  94%  |                                                                              |===================================================================   |  96%  |                                                                              |===================================================================== |  98%  |                                                                              |======================================================================| 100%

# remove (some) bias and sparsity
That <- desparsify(fit)

# de-sparsified partial correlations
That$P
#>               [,1]        [,2]        [,3]        [,4]         [,5]
#>  [1,]  0.000000000  0.07190758  0.05094874 -0.16479200 -0.156117182
#>  [2,]  0.071907577  0.00000000  0.05182416  0.27833694  0.198798358
#>  [3,]  0.050948741  0.05182416  0.00000000 -0.06457824 -0.319861820
#>  [4,] -0.164791999  0.27833694 -0.06457824  0.00000000 -0.045986586
#>  [5,] -0.156117182  0.19879836 -0.31986182 -0.04598659  0.000000000
#>  [6,]  0.002719739  0.02538559  0.48994703  0.04264428 -0.070437297
#>  [7,]  0.023912164 -0.05722995 -0.05224299  0.01747688  0.134436628
#>  [8,]  0.691616344 -0.42744543 -0.06378050  0.17301302  0.009648906
#>  [9,]  0.112201696 -0.12291158  0.07933048  0.53080816 -0.184095840
#> [10,] -0.228687138  0.12269847  0.12112253  0.26037116 -0.058537548
#> [11,] -0.033235683  0.73362264  0.11839068 -0.19364949 -0.059263154
#>               [,6]        [,7]         [,8]        [,9]       [,10]
#>  [1,]  0.002719739  0.02391216  0.691616344  0.11220170 -0.22868714
#>  [2,]  0.025385586 -0.05722995 -0.427445427 -0.12291158  0.12269847
#>  [3,]  0.489947032 -0.05224299 -0.063780502  0.07933048  0.12112253
#>  [4,]  0.042644280  0.01747688  0.173013020  0.53080816  0.26037116
#>  [5,] -0.070437297  0.13443663  0.009648906 -0.18409584 -0.05853755
#>  [6,]  0.000000000  0.40466078  0.023275912 -0.02054467 -0.01796091
#>  [7,]  0.404660777  0.00000000 -0.030363514 -0.01242277 -0.10455942
#>  [8,]  0.023275912 -0.03036351  0.000000000 -0.05247622  0.30530686
#>  [9,] -0.020544669 -0.01242277 -0.052476221  0.00000000  0.19674864
#> [10,] -0.017960913 -0.10455942  0.305306859  0.19674864  0.00000000
#> [11,] -0.009837089 -0.02495264  0.417494377  0.18691388 -0.03663081
#>              [,11]
#>  [1,] -0.033235683
#>  [2,]  0.733622641
#>  [3,]  0.118390679
#>  [4,] -0.193649492
#>  [5,] -0.059263154
#>  [6,] -0.009837089
#>  [7,] -0.024952642
#>  [8,]  0.417494377
#>  [9,]  0.186913880
#> [10,] -0.036630805
#> [11,]  0.000000000