Bootstrapped Edge Inclusion 'Probabilities'

Compute the number of times each edge was selected when performing a non-parametric bootstrap (see Figure 6.7, Hastie et al. 2009) .

boot_eip(Y, method = "pearson", samples = 500, progress = TRUE, ...)

Arguments

Y
A matrix of dimensions n by p.

method
Character string. Which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman".

samples
Numeric. How many bootstrap samples (defaults to 500)?

progress
Logical. Should a progress bar be included (defaults to TRUE)?

...
Additional arguments passed to ggmncv

Value

An object of class eip that includes the "probabilities" in a data frame.

Note

Although Hastie et al. (2009) suggests this approach provides probabilities, to avoid confusion with Bayesian inference, these are better thought of as "probabilities" (or better yet proportions).

References

Hastie T, Tibshirani R, Friedman J (2009). The elements of statistical learning: data mining, inference, and prediction. Springer Science \& Business Media.

Examples

# \donttest{
# data (ptsd symptoms)
Y <- GGMncv::ptsd[,1:10]

# compute eip's
boot_samps <- boot_eip(Y, samples  = 100, progress = FALSE)

boot_samps
#> Edge Inclusion 'Probabilities':
#> 
#>  Relation  EIP
#>    B1--B2 0.98
#>    B1--B3 0.29
#>    B2--B3 1.00
#>    B1--B4 1.00
#>    B2--B4 0.20
#>    B3--B4 0.98
#>    B1--B5 0.79
#>    B2--B5 0.24
#>    B3--B5 0.88
#>    B4--B5 0.99
#>    B1--C1 0.14
#>    B2--C1 0.04
#>    B3--C1 0.04
#>    B4--C1 0.46
#>    B5--C1 0.88
#>    B1--C2 0.08
#>    B2--C2 0.73
#>    B3--C2 0.14
#>    B4--C2 0.56
#>    B5--C2 0.91
#>    C1--C2 0.94
#>    B1--D1 0.34
#>    B2--D1 0.14
#>    B3--D1 0.13
#>    B4--D1 0.07
#>    B5--D1 0.12
#>    C1--D1 0.44
#>    C2--D1 0.38
#>    B1--D2 0.57
#>    B2--D2 0.16
#>    B3--D2 0.07
#>    B4--D2 0.37
#>    B5--D2 0.08
#>    C1--D2 0.43
#>    C2--D2 0.44
#>    D1--D2 0.54
#>    B1--D3 0.22
#>    B2--D3 0.39
#>    B3--D3 0.26
#>    B4--D3 0.85
#>    B5--D3 0.40
#>    C1--D3 0.26
#>    C2--D3 0.12
#>    D1--D3 0.65
#>    D2--D3 0.95
#> -----
# }