Confirm Edges

Confirmatory hypothesis testing of edges that were initially detected with data-driven model selection.

confirm_edges(object, Rnew, method, alpha)

Arguments

object
An object of class ggmncv

Rnew
Matrix. A correlation matrix of dimensions p by p.

method
Character string. A correction method for multiple comparison (defaults to fdr). Can be abbreviated. See p.adjust.

alpha
Numeric. Significance level (defaults to 0.05).

Value

An object of class ggmncv, including:

P: Matrix of confirmed edges (partial correlations)

adj: Matrix of confirmed edges (adjacency)

Details

The basic idea is to merge exploration with confirmation (see for example, Rodriguez et al. 2020) . This is accomplished by testing those edges (in an independent dataset) that were initially detected via data driven model selection.

Confirmatory hypothesis testing follows the approach described in Jankova and Van De Geer (2015) : (1) graphical lasso is computed with lambda fixed to \(\lambda = \sqrt{log(p)/n}\), (2) the de-sparsified estimator is computed, and then (3) p-values are obtained for the de-sparsified estimator.

References

Jankova J, Van De Geer S (2015). “Confidence intervals for high-dimensional inverse covariance estimation.” Electronic Journal of Statistics, 9(1), 1205--1229.

Rodriguez JE, Williams DR, Rast P, Mulder J (2020). “On Formalizing Theoretical Expectations: Bayesian Testing of Central Structures in Psychological Networks.” PsyArXiv. doi: 10.31234/osf.io/zw7pf .

Examples

Y <- na.omit(bfi[,1:25])

Y_explore <- Y[1:1000,]

Y_confirm <- Y[1001:nrow(Y),]

fit <- ggmncv(cor(Y_explore),
              n = nrow(Y_explore),
              progress = FALSE)

confirm <- confirm_edges(fit,
                         Rnew = cor(Y_confirm),
                         method = "fdr",
                         alpha = 0.05)