Network Comparison Test

A re-implementation and extension of the permutation based network comparison test introduced in Van Borkulo et al. (2017) . Such extensions include scaling to networks with many nodes and the option to use custom test-statistics.

nct(
  Y_g1,
  Y_g2,
  iter = 1000,
  desparsify = TRUE,
  method = "pearson",
  FUN = NULL,
  cores = 1,
  progress = TRUE,
  ...
)

Arguments

Y_g1
A matrix (or data.frame) of dimensions n by p, corresponding to the first dataset (p must be the same for Y_g1 and Y_g2).

Y_g2
A matrix of dimensions n by p, corresponding to the second dataset (p must be the same for Y_g1 and Y_g2).

iter
Numeric. Number of (Monte Carlo) permutations (defaults to 1000).

desparsify
Logical. Should the de-sparsified glasso estimator be computed (defaults to TRUE)? This is much faster, as the tuning parameter is fixed to \(\lambda = \sqrt{log(p)/n}\).

method
character string. Which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman".

FUN
A function or list of functions (defaults to NULL), specifying custom test-statistics. See Examples.

cores
Numeric. Number of cores to use when executing the permutations in parallel (defaults to 1).

...
Additional arguments passed to ggmncv.

References

Van Borkulo CD, Boschloo L, Kossakowski J, Tio P, Schoevers RA, Borsboom D, Waldorp LJ (2017). “Comparing network structures on three aspects: A permutation test.” Manuscript submitted for publication, 10.